requirements.txt
# Web Framework
flask==2.3.3
Werkzeug==2.3.7
flask-restx==1.3.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Image Processing
Pillow==10.0.0
pdf2image==1.16.3

# AI and OCR
anthropic==0.18.1

# Flight Data
amadeus==9.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
